+++ app/components/Views/Root/index.js
@@ -5,16 +5,12 @@ import { PersistGate } from 'redux-persist/lib/integration/react';
 
 import { store, persistor } from '../../../store/';
 import SplashScreen from 'react-native-splash-screen';
-import Branch from 'react-native-branch';
 
 import App from '../../Nav/App';
 import SecureKeychain from '../../../core/SecureKeychain';
 import EntryScriptWeb3 from '../../../core/EntryScriptWeb3';
 import Logger from '../../../util/Logger';
 import ErrorBoundary from '../ErrorBoundary';
-import SharedDeeplinkManager from '../../../core/DeeplinkManager';
-import Engine from '../../../core/Engine';
-import AppConstants from '../../../core/AppConstants';
 
 /**
  * Top level of the component hierarchy
@@ -22,18 +18,12 @@ import AppConstants from '../../../core/AppConstants';
  */
 export default class Root extends PureComponent {
 	static propTypes = {
-		foxCode: PropTypes.string,
-		/**
-		 * Object that represents the navigator
-		 */
-		navigation: PropTypes.object
+		foxCode: PropTypes.string
 	};
 
 	static defaultProps = {
 		foxCode: 'null'
 	};
-
-	unsubscribeFromBranch;
 
 	errorHandler = (error, stackTrace) => {
 		Logger.error(error, stackTrace);
@@ -46,33 +36,6 @@ export default class Root extends PureComponent {
 		EntryScriptWeb3.init();
 		SplashScreen.hide();
 	}
-
-	handleDeeplinks = async ({ error, params, uri }) => {
-		if (error) {
-			Logger.error(error, 'Deeplink: Error from Branch');
-		}
-		const deeplink = params['+non_branch_link'] || uri || null;
-		try {
-			if (deeplink) {
-				const { KeyringController } = Engine.context;
-				const isUnlocked = KeyringController.isUnlocked();
-				isUnlocked
-					? SharedDeeplinkManager.parse(deeplink, { origin: AppConstants.DEEPLINKS.ORIGIN_DEEPLINK })
-					: SharedDeeplinkManager.setDeeplink(deeplink);
-			}
-		} catch (e) {
-			Logger.error(e, `Deeplink: Error parsing deeplink`);
-		}
-	};
-
-	componentDidMount = () => {
-		SharedDeeplinkManager.init(this.props.navigation);
-		this.unsubscribeFromBranch = Branch.subscribe(this.handleDeeplinks);
-	};
-
-	componentWillUnmount = () => {
-		this.unsubscribeFromBranch && this.unsubscribeFromBranch();
-	};
 
 	render = () => (
 		<Provider store={store}>
